+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerAdaptors.java
@@ -1,7 +1,7 @@
 package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.HandlerExecutionAdaptor;
-import com.interface21.webmvc.servlet.mvc.controller.ControllerAdaptor;
+import com.interface21.webmvc.servlet.mvc.controller.InterfaceControllerAdaptor;
+import com.interface21.webmvc.servlet.mvc.handler.AnnotationControllerAdaptor;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -10,11 +10,11 @@ public class HandlerAdaptors {
     private final List<HandlerAdaptor> handlerAdaptors = new ArrayList<>();
 
     public HandlerAdaptors() {
-        ControllerAdaptor controllerAdaptor = new ControllerAdaptor();
-        handlerAdaptors.add(controllerAdaptor);
+        HandlerAdaptor interfaceControllerAdaptor = new InterfaceControllerAdaptor();
+        handlerAdaptors.add(interfaceControllerAdaptor);
 
-        HandlerExecutionAdaptor handlerExecutionAdaptor = new HandlerExecutionAdaptor();
-        handlerAdaptors.add(handlerExecutionAdaptor);
+        HandlerAdaptor annotationControllerAdaptor = new AnnotationControllerAdaptor();
+        handlerAdaptors.add(annotationControllerAdaptor);
     }
 
     public HandlerAdaptor findAdaptor(Object handler) {
